+++ prog/opencl_module_spinors.h
@@ -177,7 +177,7 @@ public:
 	 *
 	 * @param in Name of the kernel under consideration.
 	 */
-	virtual int get_read_write_size(const char * in);
+	virtual size_t get_read_write_size(const char * in);
 
 	/**
 	 * Return amount of Floating point operations performed by a specific kernel per call.
